requirements.txt
PyMuPDF
google-genai>=1.0.0
python-dotenv
streamlit
pandas
plotly
python-dotenv
streamlit-authenticator==0.2.1